refactor media components to standardize image loading and improve structure
