+++ presentation/designSystem/src/main/res/drawable/place_holder.png
